feat: apply succeed modal and learn more link
